--- subdivide/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "subdivide"
-edition = "2021"
-version.workspace = true
-authors.workspace = true
-
-[dependencies]
